pyproject.toml
[tool.poetry]
name = "py-api-starter"
version = "0.1.0"
description = "FastAPI + Jobs + Notebooks starter template"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.30"}
python-multipart = "^0.0.6"
sqlmodel = "^0.0.22"
psycopg = "^3.2"
alembic = "^1.13"
pydantic = {extras = ["email"], version = "^2.8"}
pydantic-settings = "^2.3"
python-jose = {extras = ["cryptography"], version = "^3.3"}
passlib = {extras = ["bcrypt"], version = "^1.7"}
stripe = "^10.0"
pandas = "^2.2"
polars = "^1.5"
jupyter = "^1.0"
apscheduler = "^3.10"
redis = "^5.0"
celery = "^5.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5"
pytest = "^8.2"
mypy = "^1.11"
pytest-asyncio = "^0.23"
httpx = "^0.27"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
